Make iter a size_t
Negative iterations don't have any meaning, so why not make it a size_t

+++ includes/fractol.h
@@ -31,7 +31,7 @@ typedef enum e_debug
 
 typedef struct s_camera
 {
-	int			iter;
+	size_t		iter;
 	long double	scale;
 	t_complex	center;
 	t_complex	top_left;
